pyproject.toml
[project]
name = "tms4_data_reader"
version = "1.0.0"
authors = [
  {name = "Erkka Rinne", email = "erkka.rinne@fmi.fi"},
]
description = "Python module for reading TOMST TMS-4 data"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
]
dependencies = [
    "python-dateutil",
    "pandas",
]

[tool.setuptools]
py-modules = ["tms4_data_reader"]
